backend/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
pandas
passlib[bcrypt]
python-jose 